==> src/03_github-user-classification/government/02_azindex-to-relational.Rmd <==
---
title: "04_gov-relational-structure"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
rm(list = ls())
# load packages 
for (pkg in c("tidyverse", "igraph", "data.table", "R.utils", "countrycode", 
              "RPostgreSQL", "cowplot", "maditr", "janitor", "ggraph")) {library(pkg, character.only = TRUE)}

# connect to postgresql to get our data
conn <- dbConnect(drv = PostgreSQL(), 
                  dbname = "sdad", 
                  host = "10.250.124.195", 
                  port = 5432, 
                  user = Sys.getenv("db_userid"), 
                  password = Sys.getenv("db_pwd"))

# query the users_gh data (table of all github users) 
us_gov_azindex <- dbGetQuery(conn, "SELECT * FROM us_gov_depts.us_gov_azindex_clean")

# disconnect from postgresql database 
dbDisconnect(conn)

```

```{r}
# load local functions 
source("/sfs/qumulo/qhome/kb7hp/oss-2020/functions/standardize_gov.R")

branch_order <- c("executive office of the u.s. president", "u.s. executive branch", "u.s. legislative branch", "u.s. judicial branch", 
                  "u.s. independent agency", "u.s. independent board, commission, committee", "u.s. quasi-official governmental institution", "none")
agency_order <- c("white house", "u.s. house of representatives", "u.s. senate", "supreme court of the u.s.",
                  "u.s. department of agriculture", "u.s. department of commerce", 
                  "u.s. department of defense", "u.s. department of education", "u.s. department of energy",
                  "u.s. department of health and human services", "u.s. department of homeland security",
                  "u.s. department of housing and urban development", "u.s. department of justice",
                  "u.s. department of labor", "u.s. department of state", "u.s. department of the interior",
                  "u.s. department of the treasury", "u.s. department of transportation", "u.s. department of veterans affairs")
#subagency_order <- c()

us_gov_azindex_dedup <- us_gov_azindex %>% 
  # clean agency
  rename(original = agency) %>% 
  standardize_gov(original) %>% 
  rename(agency_o = original, unique_institution = institution) %>% 
  # clean branch 
  rename(original = gov_branch) %>% 
  standardize_gov(original) %>% 
  dt_mutate(institution = ifelse(test = str_detect(string = institution, 
                                 pattern = "\\b(?i)(legislative)\\b"), 
                                 yes = "u.s. legislative branch", no = institution)) %>%
  dt_mutate(institution = ifelse(test = str_detect(string = institution, 
                                 pattern = "\\b(?i)(judicial)\\b"), 
                                 yes = "u.s. judicial branch", no = institution)) %>%
  dt_mutate(institution = ifelse(test = str_detect(string = institution, 
                                 pattern = "\\b(?i)(quasi-official)\\b"), 
                                 yes = "u.s. quasi-official governmental institution", no = institution)) %>%
  dt_mutate(institution = ifelse(test = str_detect(string = institution, 
                                 pattern = "\\b(?i)(none)\\b"), 
                                 yes = "u.s. independent agency", no = institution)) %>%
  rename(branch_o = original, branch = institution) %>% 
  # clean gov_agency
  rename(original = gov_agency) %>% 
  standardize_gov(original) %>% 
  rename(gov_agency_o = original, agency = institution) %>% 
  # clean child_agency
  rename(original = child_agency) %>% 
  standardize_gov(original) %>% 
  rename(child_agency_o = original, subagency = institution) %>% 
  #mutate(subagency = replace_na(subagency, "none")) %>% 
  select(branch, agency, subagency, unique_institution, everything()) %>% 
  distinct(branch, agency, subagency, unique_institution, website, other_website) %>%
  arrange(match(branch, branch_order), match(agency, agency_order), agency, subagency)

us_gov_azindex_dedup <- us_gov_azindex_dedup %>%
  filter(branch != agency | branch != subagency | branch != unique_institution |
         agency != subagency | agency != unique_institution | subagency != unique_institution) 


us_gov_azindex_dedup <- us_gov_azindex_dedup %>% 
  # no cases of branch matching agency, subagency or institution
  # four cases of agency == unique_institution
  filter(agency != unique_institution | is.na(agency)) 

us_gov_azindex_dedup <- us_gov_azindex_dedup %>% 
  # three cases of subagency == unique_institution
  filter(subagency != unique_institution | is.na(subagency)) 

us_gov_azindex_dedup
```

```{r}
# variations of this code snip were used to deduplicate by recoding variability in the original institutions 
us_gov_azindex_dedup %>% 
  #filter(agency == "u.s. agency for global media") %>% 
  get_dupes(website, branch, agency) %>% 
  select(website, unique_institution, everything())
```

### Creating the relational hierarchy 

I am cutting gov level by gov level rather than doing this all column by column. Its more code but (1) I want to validate what I see in each step and (2) I want all of the organizations standardized for use outside of this context. 

```{r}
# let's start with the white house council 
wh_br_ag <- us_gov_azindex_dedup %>% 
  filter(agency == "white house") %>% 
  select(branch, agency) %>% 
  rename(from=branch, to=agency) %>% 
  group_by(from, to) %>% 
  count() %>% rename(weight = n)

# then white house to depts 
wh_ag_inst <- us_gov_azindex_dedup %>% 
  filter(agency == "white house") %>% 
  select(agency, unique_institution) %>% 
  rename(from=agency, to=unique_institution) %>% 
  group_by(from, to) %>% 
  count() %>% rename(weight = n)

# then to the 15 executive departmnents 
executive_depts <- c("u.s. department of agriculture", "u.s. department of commerce", "u.s. department of defense",
                     "u.s. department of education", "u.s. department of energy", "u.s. department of health and human services",
                     "u.s. department of homeland security", "u.s. department of housing and urban development", "u.s. department of justice",
                     "u.s. department of labor", "u.s. department of the interior", "u.s. department of state",
                     "u.s. department of the treasury", "u.s. department of transportation", "u.s. department of veterans affairs")
exec_depts <- us_gov_azindex_dedup %>% 
  filter(branch == "u.s. executive branch" & agency %in% executive_depts) %>% 
  select(branch, agency) %>% 
  rename(from=branch, to=agency) %>% 
  group_by(from, to) %>% 
  count() %>% rename(weight = n)

# then connect the exec depts to their exec subagencies 
exec_depts_subag <- us_gov_azindex_dedup %>% 
  filter(branch == "u.s. executive branch" & agency %in% executive_depts) %>% 
  select(agency, subagency) %>% 
  rename(from=agency, to=subagency) %>% 
  drop_na(to) %>% 
  group_by(from, to) %>% 
  count() %>% rename(weight = n)

# then the exec subagencies to unique institutions  
exec_subag_inst <- us_gov_azindex_dedup %>% 
  filter(branch == "u.s. executive branch" & agency %in% executive_depts) %>% 
  select(subagency, unique_institution) %>% 
  rename(from=subagency, to=unique_institution) %>% 
  drop_na(from) %>% 
  group_by(from, to) %>% 
  count() %>% rename(weight = n)

# then the exec depts to unique institutions without subagencies
exec_ag_inst <- us_gov_azindex_dedup %>% 
  filter(branch == "u.s. executive branch" & agency %in% executive_depts) %>% 
  filter(is.na(subagency)) %>% 
  select(agency, unique_institution) %>% 
  rename(from=agency, to=unique_institution) %>% 
  group_by(from, to) %>% 
  count() %>% rename(weight = n)
 
# bind all the dfs together and plot the network 
wh_exec <- rbind(wh_br_ag, wh_ag_inst, exec_depts, exec_depts_subag, exec_subag_inst, exec_ag_inst)
wh_exec

wh_exec_g <- graph_from_data_frame(wh_exec, directed = FALSE)
E(wh_exec_g)$width <- 1+E(wh_exec_g)$weight/5 
plot(wh_exec_g, 
     weights="weight",
     layout=layout_with_fr, 
     vertex.color="skyblue", 
     vertex.size=8, 
     vertex.label=NA)
```
```{r}

# Brandon ==== Just filtered out the self-loop duplicates across the columns. The graph above shows all the WH and Exec Branch ties
# Next step is standardize all the institution names in exec_subag_inst and exec_ag_inst (starting with the natinonal institute of corrections)
# Then to map the judicial and legislation branches and then add them back into the network 
# Once the AZ index is cleaned, you need to add the scraped AZ data, the curated AZ data and the cleaned AZ data to standardizeR and share with Ledia

legjud_br_ag <- us_gov_azindex_dedup %>% 
  filter(branch %in% c("u.s. legislative branch","u.s. judicial branch")) %>% 
  rename(from=branch, to=agency) %>% 
  drop_na(to) %>% 
  group_by(from, to) %>% 
  count() %>% rename(weight = n)

legjud_ag_subag <- us_gov_azindex_dedup %>% 
  filter(branch %in% c("u.s. legislative branch","u.s. judicial branch")) %>% 
  select(agency, subagency) %>% 
  rename(from=agency, to=subagency) %>% 
  drop_na(from, to) %>% 
  group_by(from, to) %>% 
  count() %>% rename(weight = n)

legjud_subag_inst <- us_gov_azindex_dedup %>% 
  filter(branch %in% c("u.s. legislative branch","u.s. judicial branch")) %>% 
  select(subagency, unique_institution) %>% 
  rename(from=subagency, to=unique_institution) %>% 
  drop_na(from) %>% 
  group_by(from, to) %>% 
  count() %>% rename(weight = n)

legjud_ag_inst <- us_gov_azindex_dedup %>% 
  filter(branch %in% c("u.s. legislative branch","u.s. judicial branch")) %>% 
  filter(is.na(subagency)) %>% 
  select(agency, unique_institution) %>% 
  rename(from=agency, to=unique_institution) %>% 
  drop_na(from) %>% 
  group_by(from, to) %>% 
  count() %>% rename(weight = n)

# bind all the dfs together and plot the network 
all_branches <- rbind(wh_br_ag, wh_ag_inst, exec_depts, exec_depts_subag, exec_subag_inst, exec_ag_inst,
                 legjud_br_ag, legjud_ag_subag, legjud_subag_inst, legjud_ag_inst)

all_branches_g <- graph_from_data_frame(all_branches, directed = FALSE)
E(all_branches_g)$width <- 1+E(all_branches_g)$weight/5 
plot(all_branches_g, 
     weights="weight",
     layout=layout_with_fr, 
     vertex.color="skyblue", 
     vertex.size=8, 
     vertex.label=NA)
```

```{r}

# from those with agency and subagency info:
# connects branch to agency 
ind_br_ag <- us_gov_azindex_dedup %>% 
  filter(branch %in% c("u.s. independent agency", "u.s. independent board, commission, committee", "u.s. quasi-official governmental institution")) %>% 
  drop_na(agency) %>% 
  select(branch, agency) %>% 
  rename(from = branch, to = agency) %>% 
  group_by(from, to) %>% 
  count() %>% rename(weight = n)

# connects agency to subagency 
ind_ag_subag <- us_gov_azindex_dedup %>% 
  filter(branch %in% c("u.s. independent agency", "u.s. independent board, commission, committee", "u.s. quasi-official governmental institution")) %>% 
  drop_na(agency, subagency) %>% 
  select(agency, subagency) %>% 
  rename(from = agency, to = subagency) %>% 
  group_by(from, to) %>% 
  count() %>% rename(weight = n)

# connects subagency to institution 
ind_sub_inst <- us_gov_azindex_dedup %>% 
  filter(branch %in% c("u.s. independent agency", "u.s. independent board, commission, committee", "u.s. quasi-official governmental institution")) %>% 
  select(subagency, unique_institution) %>% 
  drop_na(subagency) %>% 
  rename(from = subagency, to = unique_institution) %>% 
  group_by(from, to) %>% 
  count() %>% rename(weight = n)
# and there are no branch to subagency without agency first 

# for those with no agency info
ind_br_inst <- us_gov_azindex_dedup %>% 
  filter(branch %in% c("u.s. independent agency", "u.s. independent board, commission, committee", "u.s. quasi-official governmental institution")) %>% 
  filter(is.na(agency)) %>% 
  select(branch, unique_institution) %>% 
  rename(from = branch, to = unique_institution) %>% 
  group_by(from, to) %>% 
  count() %>% rename(weight = n)
  
# bind all the dfs together and plot the network 
azindex_edgelist <- rbind(wh_br_ag, wh_ag_inst, exec_depts, exec_depts_subag, exec_subag_inst, exec_ag_inst,
                           legjud_br_ag, legjud_ag_subag, legjud_subag_inst, legjud_ag_inst, ind_br_ag, ind_ag_subag, ind_sub_inst, ind_br_inst)

azindex_graph <- graph_from_data_frame(azindex_edgelist, directed = FALSE)
E(azindex_graph)$width <- 1+E(azindex_graph)$weight/5 
plot(azindex_graph, 
     weights="weight",
     layout=layout_with_fr, 
     vertex.color="skyblue", 
     vertex.size=8, 
     vertex.label=NA)

# just have to finish up cleaning the independent agencies, etc names 

```

```{r}
# connects agency to subagency 
ind_ag_subag <- us_gov_azindex_dedup %>% 
  filter(branch %in% c("u.s. independent agency", "u.s. independent board, commission, committee", "u.s. quasi-official governmental institution")) %>% 
  drop_na(agency, subagency) %>% 
  select(agency, subagency) %>% 
  rename(from = agency, to = subagency) %>% 
  group_by(from, to) %>% 
  count() %>% rename(weight = n)

# connects subagency to institution 
ind_sub_inst <- us_gov_azindex_dedup %>% 
  filter(branch %in% c("u.s. independent agency", "u.s. independent board, commission, committee", "u.s. quasi-official governmental institution")) %>% 
  select(subagency, unique_institution) %>% 
  drop_na(subagency) %>% 
  rename(from = subagency, to = unique_institution) %>% 
  group_by(from, to) %>% 
  count() %>% rename(weight = n)
# and there are no branch to subagency without agency first 

# for those with no agency info
ind_br_inst <- us_gov_azindex_dedup %>% 
  filter(branch %in% c("u.s. independent agency", "u.s. independent board, commission, committee", "u.s. quasi-official governmental institution")) %>% 
  filter(is.na(agency)) %>% 
  select(branch, unique_institution) %>% 
  rename(from = branch, to = unique_institution) %>% 
  group_by(from, to) %>% 
  count() %>% rename(weight = n)
  
# bind all the dfs together and plot the network 
azindex_edgelist <- rbind(wh_br_ag, wh_ag_inst, exec_depts, exec_depts_subag, exec_subag_inst, exec_ag_inst,
                           legjud_br_ag, legjud_ag_subag, legjud_subag_inst, legjud_ag_inst, ind_br_ag, ind_ag_subag, ind_sub_inst, ind_br_inst)

azindex_graph <- graph_from_data_frame(azindex_edgelist, directed = FALSE)
E(azindex_graph)$width <- 1+E(azindex_graph)$weight/5 
plot(azindex_graph, 
     weights="weight",
     layout=layout_with_fr, 
     vertex.color="skyblue", 
     vertex.size=8, 
     vertex.label=NA)
```

```{r}

#### FINAL 

# now lets do the whole thing in shorter form just to make sure we don't have errors 

# branch to agency (when agency <> NA)
branch_agency <- us_gov_azindex_dedup %>% 
  drop_na(agency) %>% 
  select(branch, agency) %>% 
  rename(from = branch, to = agency) %>% 
  group_by(from, to) %>% 
  count() %>% rename(weight = n)

# branch to subagency (when agency <> NA)
# only medpac and macpac
branch_subagency <- us_gov_azindex_dedup %>% 
  filter(is.na(agency)) %>% 
  drop_na(subagency) %>% 
  select(branch, unique_institution) %>% 
  rename(from = branch, to = unique_institution) %>% 
  group_by(from, to) %>% 
  count() %>% rename(weight = n)

# agency to subagency (when agency <> NA & subagency <> NA)
agency_subagency <- us_gov_azindex_dedup %>% 
  drop_na(agency, subagency) %>% 
  select(agency, subagency) %>% 
  rename(from = agency, to = subagency) %>% 
  group_by(from, to) %>% 
  count() %>% rename(weight = n)

# branch to institution (when agency = NA)
branch_institution <- us_gov_azindex_dedup %>% 
  filter(is.na(agency)) %>% 
  select(branch, unique_institution) %>% 
  rename(from = branch, to = unique_institution) %>% 
  group_by(from, to) %>% 
  count() %>% rename(weight = n)

# agency to institution (when agency <> NA & subagency = NA)
agency_institution <- us_gov_azindex_dedup %>% 
  drop_na(agency) %>%
  filter(is.na(subagency)) %>% 
  select(agency, unique_institution) %>% 
  rename(from = agency, to = unique_institution) %>% 
  group_by(from, to) %>% 
  count() %>% rename(weight = n)

# subagency to institution 
subagency_institution <- us_gov_azindex_dedup %>% 
  drop_na(subagency) %>%
  select(subagency, unique_institution) %>% 
  rename(from = subagency, to = unique_institution) %>% 
  group_by(from, to) %>% 
  count() %>% rename(weight = n)

usgov_edgelist <- bind_rows(branch_agency, branch_subagency, agency_subagency, branch_institution, agency_institution, subagency_institution)

usgov_network <- graph_from_data_frame(usgov_edgelist, directed = FALSE)
E(usgov_network)$width <- 1+E(usgov_network)$weight/5 
plot(usgov_network, 
     weights="weight",
     layout=layout_with_fr, 
     vertex.color="skyblue", 
     vertex.size=4, 
     vertex.label=NA)
	
```

```{r}
# create a vertices data.frame. One line per object of our hierarchy, giving features of nodes.
us_gov_vertices <- data.frame(name = unique(c(as.character(usgov_edgelist$from), as.character(usgov_edgelist$to))) )
length(us_gov_vertices$name)

us_gov_vertices <- us_gov_vertices %>% 
  rename(unique_institution = name) %>% 
  left_join(us_gov_azindex_dedup, by = "unique_institution") %>% 
  arrange(unique_institution) %>% 
  select(unique_institution, website) 
```

```{r}
usgov_azindex <- us_gov_azindex_dedup %>% 
  rename(institution = unique_institution) 

write.csv(usgov_azindex, "/sfs/qumulo/qhome/kb7hp/oss-2020/data/usgov_azindex.csv", row.names = FALSE)
save(usgov_azindex, file = "/sfs/qumulo/qhome/kb7hp/oss-2020/data/usgov_azindex.rda")

```


```{r}
# load local functions 
source("/sfs/qumulo/qhome/kb7hp/oss-2020/functions/usgov_to_edgelist.R")
source("/sfs/qumulo/qhome/kb7hp/oss-2020/functions/usgov_to_nodelist.R")
source("/sfs/qumulo/qhome/kb7hp/oss-2020/functions/usgov_to_graph.R")

test_edgelist <- usgov_to_edgelist(usgov_azindex)
test_nodelist <- usgov_to_nodelist(test_edgelist)
test_graph <- usgov_to_graph(usgov_azindex)

plot(usgov_network, 
     weights="weight",
     layout=layout_with_fr, 
     vertex.color="skyblue", 
     vertex.size=4, 
     vertex.label=NA)

```


### Supplement NIH information 

https://en.wikipedia.org/wiki/List_of_institutes_and_centers_of_the_National_Institutes_of_Health

### Supplement FCC information 

https://en.wikipedia.org/wiki/Federal_Communications_Commission
