==> src/03_github-user-classification/government/05_integrate-ffs-data.Rmd <==
---
title: "05_integrate-next"
output: html_document
---

```{r setup, include=FALSE}
rm(list = ls())
# install.packages(c("tidyverse", "igraph", "visNetwork", "bc3net", "data.table", "R.utils", "RPostgreSQL", "cowplot", "maditr"))

# load packages 
for (pkg in c("tidyverse", "igraph", "data.table", "R.utils", "countrycode", "tidystringdist", "naniar",
              "RPostgreSQL", "cowplot", "maditr")) {library(pkg, character.only = TRUE)}

azindex_govman_ffrdcs <- read_csv("/sfs/qumulo/qhome/kb7hp/git/oss-2020/data/azindex_govman_ffrdcs.csv")
ffs_agencies_for_urls <- read_csv("/sfs/qumulo/qhome/kb7hp/git/oss-2020/data/FFS_gov_agencies_structured.csv")
ffs_agencies_relations <- read_csv("/sfs/qumulo/qhome/kb7hp/git/oss-2020/data/FFS_gov_agencies_structured_noloops.csv")
ffs_agencies_timeline <- read_csv("/sfs/qumulo/qhome/kb7hp/git/oss-2020/data/ffs_agencies_timeline.csv")
```

```{r}
# load local functions 
source("/sfs/qumulo/qhome/kb7hp/git/oss-2020/functions/standardize_gov.R")

ffs_clean_step1 <- ffs_agencies_relations %>% 
  rename(institution_old = institution) %>% 
  standardize_gov(agency) %>% 
  select(-agency) %>% 
  rename(agency = institution) %>% 
  standardize_gov(subagency) %>% 
  select(-subagency) %>%
  rename(subagency = institution) %>%
  standardize_gov(institution_old) %>% 
  select(agency, subagency, institution, agency_url, -institution_old) 

# to prevent duplicates i will pull out all of the unique ids from our previous aggregation of datasets and remove those from our ffs file 
unique_institutions <- unique(azindex_govman_ffrdcs$institution)
unique_institutions <- paste(c("\\b(?i)(zqx", na.omit(unique_institutions), "zqx)\\b"), collapse = "|")

ffs_clean_step2 <- ffs_clean_step1 %>% 
  filter(institution != unique_institutions)

ffs_clean_step3 <- ffs_clean_step2 %>% 
  mutate(branch = "u.s. executive branch") %>% 
  mutate(admin = NA) %>% 
  mutate(branch = ifelse(test = str_detect(string = agency, 
                                          pattern = "\\b(?i)(u.s. independent agency)\\b"), 
                        yes = "u.s. independent agency", no = branch)) %>% 
  mutate(agency = ifelse(test = str_detect(string = agency, 
                                          pattern = "\\b(?i)(u.s. independent agency)\\b"), 
                        yes = NA, no = agency)) %>% 
  select(branch, agency, subagency, admin, institution) %>% 
  mutate(az_index = 0, gov_man = 0, ffrdcs = 0, ffs = 1)

branch_order <- c("executive office of the u.s. president", "u.s. executive branch", "u.s. legislative branch", "u.s. judicial branch", 
                  "u.s. independent agency", "u.s. independent board, commission, committee", "u.s. quasi-official governmental institution", 
                  "u.s. international organizations")
agency_order <- c("white house", "u.s. congress", "u.s. house of representatives", "u.s. senate", "supreme court of the u.s.",
                  "u.s. department of agriculture", "u.s. department of commerce", 
                  "u.s. department of defense", "u.s. department of education", "u.s. department of energy",
                  "u.s. department of health and human services", "u.s. department of homeland security",
                  "u.s. department of housing and urban development", "u.s. department of justice",
                  "u.s. department of labor", "u.s. department of state", "u.s. department of the interior",
                  "u.s. department of the treasury", "u.s. department of transportation", "u.s. department of veterans affairs")

ffs_clean_step4 <- azindex_govman_ffrdcs %>% 
  select(branch, agency, subagency, admin, institution, az_index, gov_man, ffrdcs) %>% 
  mutate(ffs = 0) %>% 
  bind_rows(ffs_clean_step3) %>% 
  mutate(rm_dup = ifelse(test = str_detect(string = institution, 
                                          pattern = "\\b(?i)(u.s. economic research service)\\b"), yes = 1, no = 0)) %>% 
  filter(rm_dup == 0) %>% select(-rm_dup) %>% 
  arrange(match(branch, branch_order), match(agency, agency_order), subagency, admin, institution)

ffs_clean_step4
```

```{r}
# save the data 
write.csv(ffs_clean_step4, "/sfs/qumulo/qhome/kb7hp/git/oss-2020/data/azindex_govman_ffrdcs_ffs.csv", row.names = FALSE)

# load local functions 
source("/sfs/qumulo/qhome/kb7hp/git/oss-2020/functions/usgov_to_edgelist.R")
source("/sfs/qumulo/qhome/kb7hp/git/oss-2020/functions/usgov_to_nodelist.R")
source("/sfs/qumulo/qhome/kb7hp/git/oss-2020/functions/usgov_to_graph.R")

test_edgelist <- usgov_to_edgelist(ffs_clean_step4)

# the original usgov_to_edgelist function didn't catch admins 

subagency_admin <- ffs_clean_step4 %>%
    drop_na(subagency, admin) %>%
    select(subagency, admin) %>%
    rename(from = subagency, to = admin) %>%
    group_by(from, to) %>%
    count() %>% rename(weight = n)

admin_institution <- ffs_clean_step4 %>%
    drop_na(admin) %>%
    select(admin, institution) %>%
    rename(from = admin, to = institution) %>%
    group_by(from, to) %>%
    count() %>% rename(weight = n)

test_edgelist <- rbind(test_edgelist, subagency_admin, admin_institution)
test_nodelist <- usgov_to_nodelist(test_edgelist)
test_graph <- graph_from_data_frame(test_edgelist)
test_graph <- as.undirected(test_graph)

plot(test_graph, 
     weights="weight",
     layout=layout_with_fr, 
     vertex.color="skyblue", 
     vertex.size=4, 
     arrow.size=0,
     vertex.label=NA)
```

```{r}
# load local functions 
source("/sfs/qumulo/qhome/kb7hp/git/oss-2020/functions/standardize_gov.R")

nodelist_with_urls <- ffs_agencies_for_urls %>% 
  rename(institution_old = institution) %>% 
  standardize_gov(agency) %>% 
  select(-agency) %>% 
  rename(agency = institution) %>% 
  standardize_gov(subagency) %>% 
  select(-subagency) %>%
  rename(subagency = institution) %>%
  standardize_gov(institution_old) %>% 
  select(agency, subagency, institution, agency_url, -institution_old) %>% 
  filter(agency == subagency) %>% 
  rename(node_id = subagency) %>% 
  select(-agency, -institution); nodelist_with_urls
```

```{r}
check <- azindex_govman_ffrdcs %>%
  mutate(ffs = 0) %>% 
  select(institution, az_index, gov_man, ffrdcs, ffs, contains("website"), contains("email"), contains("phone")) %>% 
  arrange(institution) %>% 
  transform(all_websites = paste(website, other_website, gm_website, ff_website, ff_other_website, sep = "; ")) %>% 
  mutate(all_websites = str_replace_all(all_websites, "NA; ", ""),
         all_websites = str_replace_all(all_websites, "; NA", "")) %>% 
  separate(all_websites, c("first_site", "second_site", "third_site", "other_sites"), "; ", extra = "merge") 
  #select(institution, #az_index, gov_man, ffrdcs, ffs, 
  #       all_websites, everything()) %>%


# this removes many duplicate websites across the cols 
check %>% 
  tidy_stringdist(first_site, second_site, method = "jaccard") %>%   
  mutate(second_site = if_else(jaccard < 0.13, "REMOVE", second_site),
         second_site = na_if(second_site, "REMOVE")) %>% 
  select(-jaccard) %>% 
  tidy_stringdist(first_site, third_site, method = "jaccard") %>%   
  mutate(third_site = if_else(jaccard < 0.13, "REMOVE", third_site),
         third_site = na_if(third_site, "REMOVE")) %>%
  select(-jaccard) %>% 
  tidy_stringdist(second_site, third_site, method = "jaccard") %>%   
  mutate(third_site = if_else(jaccard < 0.13, "REMOVE", third_site),
         third_site = na_if(third_site, "REMOVE"))

# LEFT OFF HERE - STILL NEED TO REMOVE SOME DUPLICATE WEBSITES 

```
